Revert "Code Modularization - Beginning"
This reverts commit 9fdcd91ef0923e329a1ed4e05adaf98465c726dc.
